Force int type instead of float

+++ case.py
@@ -56,24 +56,24 @@ class case(wx.Window) :
             dc.SetFont(self.font_jeton)
             text = self._jeton.get_lettre().upper()
             l,h = dc.GetTextExtent(text)
-            dc.DrawText(text, (size_jeton - l)/2 + 1 , (size_jeton - h)/2 + 1)
+            dc.DrawText(text, int((size_jeton - l)/2 + 1) , int((size_jeton - h)/2 + 1))
             dc.SetFont(self.font_point)
             dc.SetTextForeground(s["fontcol_points"])
             text = self._jeton.get_point()
             l,h = dc.GetTextExtent(text)
-            dc.DrawText(text, size_jeton-l, size_jeton-h)
+            dc.DrawText(text, int(size_jeton-l), int(size_jeton-h))
         # ou la flèche
         elif self._fleche in (coord.HOR, coord.VER) :
             dc.SetBrush(wx.Brush(s['col_cercle']))
             dc.SetPen(wx.Pen(s['col_cercle'], 1, wx.TRANSPARENT))
-            dc.DrawCircle(self.size/2, self.size/2, self.size/2-2)
+            dc.DrawCircle(int(self.size/2), int(self.size/2), int(self.size/2-2))
             dc.SetBrush(wx.Brush(s['col_fleche']))
             dc.SetPen(wx.Pen(s['col_fleche'], 1, wx.TRANSPARENT))
             pts = [(2.0,6.5), (5,6.5), (5,8), (8.5,5), (5,2), (5,3.5), (2.0,3.5)]
             if self._fleche == coord.HOR :
-                pts_reel = [(x*self.size/10, y*self.size/10) for (x,y) in pts]
+                pts_reel = [(int(x*self.size/10), int(y*self.size/10)) for (x,y) in pts]
             elif self._fleche == coord.VER :
-                pts_reel = [(y*self.size/10, x*self.size/10) for (x,y) in pts]
+                pts_reel = [(int(y*self.size/10), int(x*self.size/10)) for (x,y) in pts]
             dc.DrawPolygon(pts_reel)
 
     def is_vide(self) :
